Button: Move button into widgets folder
